Use ZIO Config for controlling backpressure queue size.
The tests in backpressure spec were flaky as they were setting the
system property while other tests were running creating unexpected
failures (as one of the test is purposefuly inserting invalid values).

The value is now configurable using ZIO Config which allows mocking the config
provider, but have an unfortunate side effect of slightly changing the
property name as described in the changelog and docs.